Refactor piano roll mouse handling
Move all mouse handlers from the channel and selection area components to the top component.
