+++ src/components/music/SongPianoRoll.tsx
@@ -10,9 +10,19 @@ import { SplitPaneHeader } from "ui/splitpane/SplitPaneHeader";
 import l10n from "lib/helpers/l10n";
 import { RollChannel } from "./RollChannel";
 import { RollChannelGrid } from "./RollChannelGrid";
-import { ipcRenderer } from "electron";
-import trackerActions from "store/features/tracker/trackerActions";
 import { RollChannelSelectionArea } from "./RollChannelSelectionArea";
+import { clipboard, ipcRenderer } from "electron";
+import trackerActions from "store/features/tracker/trackerActions";
+import { PatternCell } from "lib/helpers/uge/song/PatternCell";
+import { cloneDeep } from "lodash";
+import clipboardActions from "store/features/clipboard/clipboardActions";
+import trackerDocumentActions from "store/features/trackerDocument/trackerDocumentActions";
+import {
+  parsePatternToClipboard,
+  parseClipboardToPattern,
+  getInstrumentListByType,
+  getInstrumentTypeByChannel,
+} from "./helpers";
 
 const CELL_SIZE = 14;
 const MAX_NOTE = 71;
@@ -160,6 +170,48 @@ const SongGridHeader = styled.div<SongGridHeaderProps>`
   `}
 `;
 
+type BlurableDOMElement = {
+  blur: () => void;
+};
+
+const clamp = (value: number, min: number, max: number) => {
+  return Math.min(max, Math.max(min, value));
+};
+
+export interface SelectionRect {
+  x: number;
+  y: number;
+  width: number;
+  height: number;
+}
+
+interface Position {
+  x: number;
+  y: number;
+}
+
+interface NoteRenderCoordinates {
+  note: number;
+  column: number;
+}
+
+const playNotePreview = (
+  song: Song,
+  channel: number,
+  note: number,
+  instrument: number
+) => {
+  const instrumentType = getInstrumentTypeByChannel(channel) || "duty";
+  const instrumentList = getInstrumentListByType(song, instrumentType);
+  ipcRenderer.send("music-data-send", {
+    action: "preview",
+    note: note,
+    type: instrumentType,
+    instrument: instrumentList[instrument],
+    square2: channel === 1,
+  });
+};
+
 export const SongPianoRoll = ({
   song,
   sequenceId,
@@ -174,6 +226,7 @@ export const SongPianoRoll = ({
   );
 
   const patternId = song?.sequence[sequenceId] || 0;
+  const pattern = song?.patterns[patternId];
 
   const [playbackState, setPlaybackState] = useState([0, 0]);
 
@@ -233,6 +286,629 @@ export const SongPianoRoll = ({
   const visibleChannels = useSelector(
     (state: RootState) => state.tracker.visibleChannels
   );
+
+  const tool = useSelector((state: RootState) => state.tracker.tool);
+  const gridRef = useRef<HTMLDivElement>(null);
+
+  const [draggingSelection, setDraggingSelection] = useState(false);
+  const [selectionOrigin, setSelectionOrigin] =
+    useState<Position | undefined>();
+  const [selectionRect, setSelectionRect] =
+    useState<SelectionRect | undefined>();
+  const [addToSelection, setAddToSelection] = useState(false);
+
+  const selectedPatternCells = useSelector(
+    (state: RootState) => state.tracker.selectedPatternCells
+  );
+
+  useEffect(() => {
+    setRenderPattern(pattern);
+  }, [pattern]);
+
+  useEffect(() => {
+    setRenderSelectedPatternCells(selectedPatternCells);
+  }, [selectedPatternCells]);
+
+  const [isMouseDown, setIsMouseDown] = useState<boolean>(false);
+  const [isDragging, setIsDragging] = useState<boolean>(false);
+  const [clonePatternCells, setClonePatternCells] = useState<boolean>(false);
+
+  const [moveNoteFrom, setMoveNoteFrom] = useState<NoteRenderCoordinates>();
+  const [moveNoteTo, setMoveNoteTo] = useState<NoteRenderCoordinates>();
+  const [renderPattern, setRenderPattern] = useState<PatternCell[][]>();
+  const [renderSelectedPatternCells, setRenderSelectedPatternCells] = useState<
+    number[]
+  >([]);
+
+  const [pastedPattern, setPastedPattern] = useState<PatternCell[][]>();
+
+  const hoverColumn = useSelector(
+    (state: RootState) => state.tracker.hoverColumn
+  );
+
+  const defaultInstruments = useSelector(
+    (state: RootState) => state.tracker.defaultInstruments
+  );
+  const currentInstrument = defaultInstruments[selectedChannel];
+
+  useEffect(() => {
+    if (tool !== "selection") {
+      setDraggingSelection(false);
+      setSelectionRect(undefined);
+    }
+  }, [tool]);
+
+  const selectCellsInRange = useCallback(
+    (selectedPatternCells: number[], selectionRect: SelectionRect) => {
+      if (pattern) {
+        const fromI = selectionRect.x / CELL_SIZE;
+        const toI = (selectionRect.x + selectionRect.width) / CELL_SIZE;
+        const fromJ = selectionRect.y / CELL_SIZE;
+        const toJ = (selectionRect.y + selectionRect.height) / CELL_SIZE;
+
+        const newSelectedPatterns = [...selectedPatternCells];
+        for (let i = fromI; i < toI; i++) {
+          for (let j = fromJ; j < toJ; j++) {
+            const note = 12 * 6 - 1 - j;
+            if (
+              pattern[i][selectedChannel] &&
+              pattern[i][selectedChannel].note === note
+            ) {
+              newSelectedPatterns.push(i);
+            }
+          }
+        }
+        return newSelectedPatterns;
+      }
+      return [];
+    },
+    [selectedChannel, pattern]
+  );
+
+  const onSelectAll = useCallback(
+    (_e) => {
+      const selection = window.getSelection();
+      if (!selection || selection.focusNode) {
+        return;
+      }
+      window.getSelection()?.empty();
+      const allPatternCells = pattern
+        ?.map((c, i) => {
+          return c[selectedChannel].note !== null ? i : undefined;
+        })
+        .filter((c) => c !== undefined) as number[];
+      dispatch(trackerActions.setSelectedPatternCells(allPatternCells));
+
+      // Blur any focused element to be able to use keyboard actions on the
+      // selection
+      const el = document.querySelector(":focus") as unknown as
+        | BlurableDOMElement
+        | undefined;
+      if (el && el.blur) el.blur();
+    },
+    [selectedChannel, dispatch, pattern]
+  );
+
+  useEffect(() => {
+    document.addEventListener("selectionchange", onSelectAll);
+    return () => {
+      document.removeEventListener("selectionchange", onSelectAll);
+    };
+  }, [onSelectAll]);
+
+  const refreshRenderPattern = useCallback(
+    (newMoveNoteTo) => {
+      if (pattern) {
+        const columnFrom = moveNoteFrom?.column || 0;
+        const columnTo = newMoveNoteTo?.column || 0;
+        const columnDir = columnFrom > columnTo ? -1 : 1;
+        const deltaX = Math.abs(columnFrom - columnTo) * columnDir;
+
+        const noteFrom = moveNoteFrom?.note || 0;
+        const noteTo = newMoveNoteTo?.note || 0;
+        const noteDir = noteFrom > noteTo ? -1 : 1;
+        const deltaY = Math.abs(noteFrom - noteTo) * noteDir;
+
+        const newPattern = cloneDeep(pattern);
+        for (const i of selectedPatternCells) {
+          const newPatternColumn = cloneDeep(pattern[i]);
+          const newPatternCell = !clonePatternCells
+            ? newPatternColumn.splice(selectedChannel, 1, new PatternCell())[0]
+            : { ...newPatternColumn[selectedChannel] };
+          if (newPattern[i + deltaX]) {
+            newPatternCell.note =
+              newPatternCell.note !== null
+                ? (newPatternCell.note + deltaY + 12 * 6) % (12 * 6)
+                : null;
+            if (selectedPatternCells.indexOf(i - deltaX) === -1) {
+              newPattern[i] = newPatternColumn;
+            }
+            newPattern[i + deltaX][selectedChannel] = newPatternCell;
+          } else if (i + deltaX < 0 || i + deltaX >= 64) {
+            if (selectedPatternCells.indexOf(i - deltaX) === -1) {
+              newPattern[i][selectedChannel] = new PatternCell();
+            }
+          }
+        }
+        setRenderPattern(newPattern);
+        setRenderSelectedPatternCells(
+          selectedPatternCells.map((i) => i + deltaX)
+        );
+      }
+    },
+    [
+      selectedChannel,
+      clonePatternCells,
+      moveNoteFrom?.column,
+      moveNoteFrom?.note,
+      pattern,
+      selectedPatternCells,
+    ]
+  );
+
+  useEffect(() => {
+    if (isDragging) {
+      refreshRenderPattern(moveNoteTo);
+    }
+  }, [isDragging, moveNoteTo, refreshRenderPattern]);
+
+  const refreshPastedPattern = useCallback(
+    (currentPastedPattern: PatternCell[][]) => {
+      if (pattern && hoverColumn !== null && hoverNote !== null) {
+        const newPattern = cloneDeep(pattern);
+
+        let columnOffset = 0;
+        let noteOffset = undefined;
+        for (const p of currentPastedPattern) {
+          const pastedPatternCell = { ...p[0] };
+          if (pastedPatternCell.note !== null) {
+            if (noteOffset === undefined) {
+              noteOffset = hoverNote - pastedPatternCell.note;
+            }
+            if (
+              hoverColumn + columnOffset >= 0 &&
+              hoverColumn + columnOffset < 64
+            ) {
+              pastedPatternCell.note =
+                ((pastedPatternCell.note + noteOffset || 0) + 12 * 6) %
+                (12 * 6);
+              newPattern[(hoverColumn + columnOffset) % 64][selectedChannel] =
+                pastedPatternCell;
+            }
+          }
+          columnOffset++;
+        }
+        setRenderPattern(newPattern);
+      }
+    },
+    [selectedChannel, hoverColumn, hoverNote, pattern]
+  );
+
+  // Mouse
+  const handleMouseDown = useCallback(
+    (e: any) => {
+      if (!pattern) return;
+      const col = Math.floor(e.offsetX / CELL_SIZE);
+      const note = 12 * 6 - 1 - Math.floor(e.offsetY / CELL_SIZE);
+      const cell = pattern[col][selectedChannel];
+
+      if (pastedPattern && renderPattern) {
+        dispatch(
+          trackerDocumentActions.editPattern({
+            patternId: patternId,
+            pattern: renderPattern,
+          })
+        );
+        setPastedPattern(undefined);
+      } else if (tool === "pencil" && e.button === 0) {
+        // If there's a note in position
+        if (cell && cell.note === note) {
+          setIsMouseDown(true);
+          setIsDragging(false);
+          setMoveNoteFrom({ column: col, note: cell.note });
+          setMoveNoteTo({ column: col, note: cell.note });
+        }
+        if (cell && cell.note !== note && selectedPatternCells.length > 1) {
+          dispatch(trackerActions.setSelectedPatternCells([]));
+        } else {
+          const changes = {
+            instrument: defaultInstruments[selectedChannel],
+            note: note,
+          };
+          dispatch(
+            trackerDocumentActions.editPatternCell({
+              patternId: patternId,
+              cell: [col, selectedChannel],
+              changes: changes,
+            })
+          );
+
+          if (song) {
+            playNotePreview(song, selectedChannel, note, currentInstrument);
+          }
+
+          if (!selectedPatternCells.includes(col)) {
+            dispatch(trackerActions.setSelectedPatternCells([col]));
+          }
+        }
+        // setMoveNoteFrom(undefined);
+      } else if (e.button === 2 || (tool === "eraser" && e.button === 0)) {
+        dispatch(
+          trackerDocumentActions.editPatternCell({
+            patternId: patternId,
+            cell: [col, selectedChannel],
+            changes: {
+              instrument: null,
+              note: null,
+            },
+          })
+        );
+        const newSelectedCells = [...selectedPatternCells];
+        newSelectedCells.splice(col, 1);
+        dispatch(trackerActions.setSelectedPatternCells(newSelectedCells));
+      } else if (tool === "selection" && e.button === 0) {
+        // If there's a note in position
+        if (cell && cell.note === note) {
+          if (!selectedPatternCells.includes(col)) {
+            if (addToSelection) {
+              const newSelectedPatterns = [...selectedPatternCells];
+              newSelectedPatterns.push(col);
+              dispatch(
+                trackerActions.setSelectedPatternCells(newSelectedPatterns)
+              );
+            } else {
+              dispatch(trackerActions.setSelectedPatternCells([col]));
+            }
+          }
+          setIsMouseDown(true);
+          setIsDragging(false);
+          setMoveNoteFrom({ column: col, note: cell.note });
+          setMoveNoteTo({ column: col, note: cell.note });
+        } else if (gridRef.current) {
+          const bounds = gridRef.current.getBoundingClientRect();
+          const x = clamp(
+            Math.floor((e.pageX - bounds.left) / CELL_SIZE) * CELL_SIZE,
+            0,
+            63 * CELL_SIZE
+          );
+          const y = clamp(
+            Math.floor((e.pageY - bounds.top) / CELL_SIZE) * CELL_SIZE,
+            0,
+            12 * 6 * CELL_SIZE - CELL_SIZE
+          );
+
+          const newSelectionRect = {
+            x,
+            y,
+            width: CELL_SIZE,
+            height: CELL_SIZE,
+          };
+
+          const newSelectedPatterns = selectCellsInRange(
+            addToSelection ? selectedPatternCells : [],
+            newSelectionRect
+          );
+
+          setSelectionOrigin({ x, y });
+          setSelectionRect(newSelectionRect);
+          setDraggingSelection(true);
+          dispatch(trackerActions.setSelectedPatternCells(newSelectedPatterns));
+        }
+      }
+    },
+    [
+      tool,
+      pastedPattern,
+      renderPattern,
+      selectedPatternCells,
+      pattern,
+      dispatch,
+      selectedChannel,
+      defaultInstruments,
+      patternId,
+      song,
+      currentInstrument,
+      selectCellsInRange,
+      addToSelection,
+    ]
+  );
+
+  const handleMouseUp = useCallback(
+    (_e: MouseEvent) => {
+      if (isDragging && isMouseDown && selectedPatternCells.length > 0) {
+        if (pattern && renderPattern) {
+          dispatch(
+            trackerDocumentActions.editPattern({
+              patternId: patternId,
+              pattern: renderPattern,
+            })
+          );
+          dispatch(
+            trackerActions.setSelectedPatternCells(
+              renderSelectedPatternCells.filter((c) => c >= 0 && c < 64)
+            )
+          );
+        }
+        setMoveNoteFrom({ note: 0, column: 0 });
+        setMoveNoteTo({ note: 0, column: 0 });
+      }
+      setSelectionRect(undefined);
+      setIsDragging(false);
+      setIsMouseDown(false);
+    },
+    [
+      dispatch,
+      isDragging,
+      isMouseDown,
+      pattern,
+      patternId,
+      renderPattern,
+      renderSelectedPatternCells,
+      selectedPatternCells.length,
+    ]
+  );
+
+  const handleMouseMove = useCallback(
+    (e: any) => {
+      const newNote = 12 * 6 - 1 - Math.floor(e.offsetY / CELL_SIZE);
+      const newColumn = Math.floor(e.offsetX / CELL_SIZE);
+
+      if (newNote !== hoverNote) {
+        dispatch(trackerActions.setHoverNote(newNote));
+      }
+
+      if (newColumn !== hoverColumn) {
+        dispatch(trackerActions.setHoverColumn(newColumn));
+      }
+
+      if (pattern && pastedPattern) {
+        refreshPastedPattern(pastedPattern);
+      } else if (isMouseDown && song && selectedPatternCells.length > 0) {
+        setIsDragging(true);
+        if (moveNoteTo?.note !== newNote || moveNoteTo?.column !== newColumn) {
+          playNotePreview(song, selectedChannel, newNote, currentInstrument);
+
+          console.log(moveNoteTo);
+          const newMoveNoteTo = {
+            note: newNote,
+            column: newColumn,
+          };
+          setMoveNoteTo(newMoveNoteTo);
+
+          refreshRenderPattern(newMoveNoteTo);
+        }
+      } else if (
+        tool === "selection" &&
+        draggingSelection &&
+        selectionRect &&
+        gridRef.current &&
+        selectionOrigin
+      ) {
+        const bounds = gridRef.current.getBoundingClientRect();
+        const x2 = clamp(
+          Math.floor((e.pageX - bounds.left) / CELL_SIZE) * CELL_SIZE,
+          0,
+          64 * CELL_SIZE
+        );
+        const y2 = clamp(
+          Math.floor((e.pageY - bounds.top) / CELL_SIZE) * CELL_SIZE,
+          0,
+          12 * 6 * CELL_SIZE
+        );
+
+        const x = Math.min(selectionOrigin.x, x2);
+        const y = Math.min(selectionOrigin.y, y2);
+        const width = Math.abs(selectionOrigin.x - x2);
+        const height = Math.abs(selectionOrigin.y - y2);
+
+        setSelectionRect({ x, y, width, height });
+
+        const selectedCells = selectCellsInRange(
+          addToSelection ? selectedPatternCells : [],
+          selectionRect
+        );
+        dispatch(trackerActions.setSelectedPatternCells(selectedCells));
+      }
+    },
+    [
+      addToSelection,
+      currentInstrument,
+      dispatch,
+      draggingSelection,
+      hoverColumn,
+      hoverNote,
+      isMouseDown,
+      moveNoteTo,
+      pastedPattern,
+      pattern,
+      refreshPastedPattern,
+      refreshRenderPattern,
+      selectCellsInRange,
+      selectedChannel,
+      selectedPatternCells,
+      selectionOrigin,
+      selectionRect,
+      song,
+      tool,
+    ]
+  );
+
+  const handleMouseLeave = useCallback(
+    (_e: MouseEvent) => {
+      if (hoverNote) {
+        dispatch(trackerActions.setHoverNote(null));
+      }
+    },
+    [hoverNote, dispatch]
+  );
+
+  useEffect(() => {
+    window.addEventListener("mousemove", handleMouseMove);
+    window.addEventListener("mouseup", handleMouseUp);
+    return () => {
+      window.removeEventListener("mousemove", handleMouseMove);
+      window.removeEventListener("mouseup", handleMouseUp);
+    };
+  }, [handleMouseMove, handleMouseUp]);
+
+  // Keyboard
+  const onKeyDown = useCallback(
+    (e: KeyboardEvent) => {
+      if ((e.target as HTMLElement)?.nodeName !== "BODY") {
+        return;
+      }
+      if (e.ctrlKey || e.metaKey) {
+        return;
+      }
+      if (e.altKey) {
+        setClonePatternCells(true);
+      }
+      if (e.shiftKey) {
+        setAddToSelection(true);
+      }
+
+      if (e.key === "Backspace") {
+        if (pattern) {
+          const newPattern = cloneDeep(pattern);
+          console.log(newPattern, selectedPatternCells, selectedChannel);
+          selectedPatternCells.forEach((i) => {
+            console.log(newPattern[i]);
+            newPattern[i].splice(selectedChannel, 1, new PatternCell());
+          });
+          dispatch(trackerActions.setSelectedPatternCells([]));
+          console.log(patternId, newPattern);
+          dispatch(
+            trackerDocumentActions.editPattern({
+              patternId: patternId,
+              pattern: newPattern,
+            })
+          );
+        }
+      }
+      if (e.key === "Escape") {
+        dispatch(trackerActions.setSelectedPatternCells([]));
+        setIsDragging(false);
+        if (pastedPattern !== undefined) {
+          setPastedPattern(undefined);
+          setRenderPattern(pattern);
+        }
+      }
+    },
+    [
+      selectedChannel,
+      dispatch,
+      pastedPattern,
+      pattern,
+      patternId,
+      selectedPatternCells,
+    ]
+  );
+
+  const onKeyUp = useCallback((e: KeyboardEvent) => {
+    if (!e.altKey) {
+      setClonePatternCells(false);
+    }
+    if (!e.shiftKey) {
+      setAddToSelection(false);
+    }
+  }, []);
+
+  useEffect(() => {
+    window.addEventListener("keydown", onKeyDown);
+    window.addEventListener("keyup", onKeyUp);
+    return () => {
+      window.removeEventListener("keydown", onKeyDown);
+      window.removeEventListener("keyup", onKeyUp);
+    };
+  }, [onKeyDown, onKeyUp]);
+
+  // Clipoard
+  const onCopy = useCallback(() => {
+    if (pattern) {
+      const parsedSelectedPattern = parsePatternToClipboard(
+        pattern,
+        selectedChannel,
+        selectedPatternCells
+      );
+      dispatch(clipboardActions.copyText(parsedSelectedPattern));
+    }
+  }, [selectedChannel, dispatch, pattern, selectedPatternCells]);
+
+  const onPaste = useCallback(() => {
+    if (pattern) {
+      const newPastedPattern = parseClipboardToPattern(clipboard.readText());
+      if (newPastedPattern) {
+        refreshPastedPattern(newPastedPattern);
+      }
+      setPastedPattern(newPastedPattern);
+      dispatch(trackerActions.setSelectedPatternCells([]));
+
+      // Blur any focused element to be able to use keyboard actions on the
+      // selection
+      const el = document.querySelector(":focus") as unknown as
+        | BlurableDOMElement
+        | undefined;
+      if (el && el.blur) el.blur();
+    }
+  }, [dispatch, pattern, refreshPastedPattern]);
+
+  const onPasteInPlace = useCallback(() => {
+    if (pattern) {
+      const newPastedPattern = parseClipboardToPattern(clipboard.readText());
+
+      if (newPastedPattern) {
+        const newPattern = Array(64)
+          .fill([
+            new PatternCell(),
+            new PatternCell(),
+            new PatternCell(),
+            new PatternCell(),
+          ])
+          .map((r, i) => {
+            console.log(i, selectedChannel);
+            const row = [...pattern[i]];
+            const pastedRow = newPastedPattern[i];
+            if (pastedRow) {
+              if (pastedRow.length === 4) {
+                return pastedRow;
+              } else {
+                pastedRow.forEach((c, i) => {
+                  row[(selectedChannel + i) % 4] = c;
+                });
+              }
+              return row;
+            }
+            return r;
+          });
+
+        console.log(newPattern);
+        dispatch(
+          trackerDocumentActions.editPattern({
+            patternId: patternId,
+            pattern: newPattern,
+          })
+        );
+      }
+
+      // Blur any focused element to be able to use keyboard actions on the
+      // selection
+      const el = document.querySelector(":focus") as unknown as
+        | BlurableDOMElement
+        | undefined;
+      if (el && el.blur) el.blur();
+    }
+  }, [selectedChannel, dispatch, pattern, patternId]);
+
+  useEffect(() => {
+    window.addEventListener("copy", onCopy);
+    window.addEventListener("paste", onPaste);
+    ipcRenderer.on("paste-in-place", onPasteInPlace);
+    return () => {
+      window.removeEventListener("copy", onCopy);
+      window.removeEventListener("paste", onPaste);
+      ipcRenderer.removeListener("paste-in-place", onPasteInPlace);
+    };
+  }, [onCopy, onPaste, onPasteInPlace]);
 
   const v = [
     selectedChannel,
@@ -345,22 +1021,39 @@ export const SongPianoRoll = ({
                   </>
                 ))}
             </Piano>
-            <SongGrid>
+            <SongGrid
+              onMouseDown={(e) => {
+                handleMouseDown(e.nativeEvent);
+              }}
+              onMouseLeave={(e) => {
+                handleMouseLeave(e.nativeEvent);
+              }}
+              // onMouseMove={(e) => {
+              //   handleMouseMove(e.nativeEvent);
+              // }}
+              style={{
+                cursor: isDragging
+                  ? clonePatternCells
+                    ? "copy"
+                    : "move"
+                  : "auto",
+              }}
+            >
               <RollChannelGrid cellSize={CELL_SIZE} />
               {v.map((i) => (
                 <RollChannel
                   channelId={i}
                   active={selectedChannel === i}
-                  patternId={patternId}
-                  pattern={song?.patterns[patternId]}
+                  renderPattern={renderPattern || []}
+                  renderSelectedPatternCells={renderSelectedPatternCells}
+                  isDragging={isDragging}
                   cellSize={CELL_SIZE}
                 />
               ))}
               <RollChannelSelectionArea
-                channelId={selectedChannel}
-                patternId={patternId}
-                pattern={song?.patterns[patternId]}
+                ref={gridRef}
                 cellSize={CELL_SIZE}
+                selectionRect={selectionRect}
               />
             </SongGrid>
           </div>
